Unsigned data type index

+++ brian2/codegen/runtime/cython_rt/templates/group_get_indices.pyx
@@ -6,7 +6,7 @@
 
     _vectorisation_idx = 1
 
-    cdef int _num_elements = 0
+    cdef size_t _num_elements = 0
     cdef _numpy.ndarray[int, ndim=1, mode='c'] _elements = _numpy.zeros(N, dtype=_numpy.int32)
     cdef int[:] _elements_view = _elements
