+++ brian2/codegen/runtime/cython_rt/templates/group_variable_get.pyx
@@ -13,7 +13,7 @@
 
     _vectorisation_idx = 1
     
-    cdef int _num_elements = 0
+    cdef size_t _num_elements = 0
     cdef _numpy.ndarray[{{c_type}}, ndim=1, mode='c'] _elements = _numpy.zeros(_num{{_group_idx}}, dtype=_numpy.{{np_type}})
     cdef {{c_type}}[:] _elements_view = _elements
