+++ brian2/codegen/runtime/cython_rt/templates/reset.pyx
@@ -7,7 +7,7 @@
     {#  Get the name of the array that stores these events (e.g. the spikespace array) #}
     {% set _eventspace = get_array_name(eventspace_variable) %}
 
-    cdef int _num_events = {{_eventspace}}[_num{{_eventspace}}-1]
+    cdef size_t _num_events = {{_eventspace}}[_num{{_eventspace}}-1]
     for _index_events in range(_num_events):
         # vector code
         _idx = {{_eventspace}}[_index_events]
