+++ brian2/codegen/runtime/weave_rt/templates/group_variable_get_conditional.cpp
@@ -5,24 +5,24 @@
     //// MAIN CODE ////////////
     // This allows everything to work correctly for synapses where N is not a
     // constant
-    const int _N = {{constant_or_scalar('N', variables['N'])}};
+    const size_t _N = {{constant_or_scalar('N', variables['N'])}};
 
     {%set c_type = c_data_type(variables['_variable'].dtype) %}
     {%set numpy_dtype = dtype(variables['_variable'].dtype).char %}
     {%set numpy_type_int = dtype(variables['_variable'].dtype).num %}
     // {{c_type}} {{numpy_dtype}} {{numpy_type_int}}
-    int _cpp_numelements = 0;
+    size_t _cpp_numelements = 0;
     // Container for all the potential values
     {{c_type}}* _elements = ({{c_type}}*)malloc(sizeof({{c_type}}) * _N);
 
     // scalar code
-	const int _vectorisation_idx = 1;
-	{{scalar_code|autoindent}}
+    const size_t _vectorisation_idx = 1;
+    {{scalar_code|autoindent}}
 
-    for(int _idx=0; _idx<_N; _idx++)
+    for(size_t _idx=0; _idx<_N; _idx++)
     {
         // vector code
-        const int _vectorisation_idx = _idx;
+        const size_t _vectorisation_idx = _idx;
         {{ super() }}
         if(_cond) {
             // _variable is set in the abstract code, see Group._get_with_code
