+++ brian2/codegen/runtime/weave_rt/templates/statemonitor.cpp
@@ -4,7 +4,7 @@
     {# USES_VARIABLES { t, _clock_t, _indices, N } #}
 
     // Get the current length and new length of t and value arrays
-    const int _new_len = {{N}} + 1;
+    const size_t _new_len = {{N}} + 1;
 
     // Resize the recorded times and get the (potentially changed) reference to
     // the underlying data
@@ -14,8 +14,8 @@
 
 
     // scalar code
-	const int _vectorisation_idx = 1;
-	{{scalar_code|autoindent}}
+    const size_t _vectorisation_idx = 1;
+    {{scalar_code|autoindent}}
 
     {% for varname, var in _recorded_variables | dictsort %}
     {%set c_type = c_data_type(variables[varname].dtype) %}
@@ -25,11 +25,11 @@
         PyObject_CallMethod(_var_{{varname}}, "resize", "((ii))", _new_len, _num_indices);
         PyArrayObject *_record_data = (((PyArrayObject*)(PyObject*){{get_array_name(var, access_data=False)}}.attr("data")));
         const npy_intp* _record_strides = _record_data->strides;
-        for (int _i = 0; _i < _num_indices; _i++)
+        for (size_t _i = 0; _i < _num_indices; _i++)
         {
             // vector code
-            const int _idx = {{_indices}}[_i];
-            const int _vectorisation_idx = _idx;
+            const size_t _idx = {{_indices}}[_i];
+            const size_t _vectorisation_idx = _idx;
             {{ super() }}
 
             {{c_type}} *recorded_entry = ({{c_type}}*)(_record_data->data + (_new_len - 1)*_record_strides[0] + _i*_record_strides[1]);
