+++ brian2/codegen/runtime/weave_rt/templates/stateupdate.cpp
@@ -3,19 +3,19 @@
 {# ALLOWS_SCALAR_WRITE #}
 
 {% block maincode %}
-	//// MAIN CODE ////////////
+    //// MAIN CODE ////////////
 
     // This allows everything to work correctly for synapses where N is not a
     // constant
-    const int _N = {{constant_or_scalar('N', variables['N'])}};
-	// scalar code
-	const int _vectorisation_idx = 1;
-	{{scalar_code|autoindent}}
+    const size_t _N = {{constant_or_scalar('N', variables['N'])}};
+    // scalar code
+    const size_t _vectorisation_idx = 1;
+    {{scalar_code|autoindent}}
 
-	for(int _idx=0; _idx<_N; _idx++)
-	{
-	    // vector code
-		const int _vectorisation_idx = _idx;
-		{{ super() }}
-	}
+    for(size_t _idx=0; _idx<_N; _idx++)
+    {
+        // vector code
+        const size_t _vectorisation_idx = _idx;
+        {{ super() }}
+    }
 {% endblock %}
